Fix: CreateOrderCommand -> TelegramUserId validation error

+++ Atlas.Application/CQRS/Orders/Commands/CreateOrder/CreateOrderCommand.cs
@@ -31,7 +31,7 @@ namespace Atlas.Application.CQRS.Orders.Commands.CreateOrder
 
         public DateTime? DeliverAt { get; set; }
 
-        public int? TelegramUserId { get; set; }
+        public long? TelegramUserId { get; set; }
 
         public bool IsDevVersionBot { get; set; }
