+++ Atlas.Domain/Order.cs
@@ -49,7 +49,7 @@ namespace Atlas.Domain
 
         public bool CanRefund { get; set; }
 
-        public int? TelegramUserId { get; set; }
+        public long? TelegramUserId { get; set; }
 
         public bool IsDevVersionBot { get; set; }
